General Improvements
Changed behav model portlist to add a pin for software mode select. Will
throw and error if hardware mode is selected.

Added internal logic to calculate Vref voltage used as outlined in
section 9.3.1.7 of the datasheet. This logic is based on the settings
specified in the internal registers.

Added print statements to show the register configurations

config.sv has options to display messages. Helper function to log is
also added.